scanfAge 추가

+++ 09_scanfChar/scanfChar.c
@@ -8,7 +8,7 @@ int main(void)
 	//printf("입력한 문자는 %c 입니다.\n", ch);
 
 	char szBuffer[4] = { 0 };
-	scanf_s("%4c", szBuffer, _countof(szBuffer));
+	scanf_s("%4c", szBuffer, (unsigned int)_countof(szBuffer));
 	printf("%c%c%c%c\n",
 		szBuffer[0], szBuffer[1], szBuffer[2], szBuffer[3]);
 	return 0;
